Fix signedness warning in gdk-pixbuf/gdk-pixbuf-loader.c
gdk-pixbuf/gdk-pixbuf-loader.c: In function ‘gdk_pixbuf_loader_eat_header_write’:
/usr/include/glib-2.0/glib/gmacros.h:336:26: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                          ^
gdk-pixbuf/gdk-pixbuf-loader.c:482:19: note: in expansion of macro ‘MIN’
         n_bytes = MIN(SNIFF_BUFFER_SIZE - priv->header_buf_offset, count);
                   ^~~
/usr/include/glib-2.0/glib/gmacros.h:336:35: error: operand of ?: changes signedness from ‘gint’ {aka ‘int’} to ‘gsize’ {aka ‘long unsigned int’} due to unsignedness of other operand [-Werror=sign-compare]
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
                                   ^~~
gdk-pixbuf/gdk-pixbuf-loader.c:482:19: note: in expansion of macro ‘MIN’
         n_bytes = MIN(SNIFF_BUFFER_SIZE - priv->header_buf_offset, count);
                   ^~~

+++ gdk-pixbuf/gdk-pixbuf-loader.c
@@ -101,7 +101,7 @@ typedef struct
         GdkPixbufAnimation *animation;
         gboolean closed;
         guchar header_buf[SNIFF_BUFFER_SIZE];
-        gint header_buf_offset;
+        gsize header_buf_offset;
         GdkPixbufModule *image_module;
         gpointer context;
         gint original_width;
